feat: modularising eth2
