+++ common/eth2/src/lighthouse.rs
@@ -221,10 +221,10 @@ impl From<Eth1Block> for FinalizedExecutionBlock {
 impl BeaconNodeHttpClient {
     /// `GET lighthouse/health`
     pub async fn get_lighthouse_health(&self) -> Result<GenericResponse<Health>, Error> {
-        let mut path = self.server.full.clone();
+        let mut path = self.server().full.clone();
 
         path.path_segments_mut()
-            .map_err(|()| Error::InvalidUrl(self.server.clone()))?
+            .map_err(|()| Error::InvalidUrl(self.server().clone()))?
             .push("lighthouse")
             .push("health");
 
@@ -233,10 +233,10 @@ impl BeaconNodeHttpClient {
 
     /// `GET lighthouse/syncing`
     pub async fn get_lighthouse_syncing(&self) -> Result<GenericResponse<SyncState>, Error> {
-        let mut path = self.server.full.clone();
+        let mut path = self.server().full.clone();
 
         path.path_segments_mut()
-            .map_err(|()| Error::InvalidUrl(self.server.clone()))?
+            .map_err(|()| Error::InvalidUrl(self.server().clone()))?
             .push("lighthouse")
             .push("syncing");
 
@@ -253,10 +253,10 @@ impl BeaconNodeHttpClient {
 
     /// `GET lighthouse/proto_array`
     pub async fn get_lighthouse_proto_array(&self) -> Result<GenericResponse<ProtoArray>, Error> {
-        let mut path = self.server.full.clone();
+        let mut path = self.server().full.clone();
 
         path.path_segments_mut()
-            .map_err(|()| Error::InvalidUrl(self.server.clone()))?
+            .map_err(|()| Error::InvalidUrl(self.server().clone()))?
             .push("lighthouse")
             .push("proto_array");
 
@@ -268,10 +268,10 @@ impl BeaconNodeHttpClient {
         &self,
         epoch: Epoch,
     ) -> Result<GenericResponse<GlobalValidatorInclusionData>, Error> {
-        let mut path = self.server.full.clone();
+        let mut path = self.server().full.clone();
 
         path.path_segments_mut()
-            .map_err(|()| Error::InvalidUrl(self.server.clone()))?
+            .map_err(|()| Error::InvalidUrl(self.server().clone()))?
             .push("lighthouse")
             .push("validator_inclusion")
             .push(&epoch.to_string())
@@ -286,10 +286,10 @@ impl BeaconNodeHttpClient {
         epoch: Epoch,
         validator_id: ValidatorId,
     ) -> Result<GenericResponse<Option<ValidatorInclusionData>>, Error> {
-        let mut path = self.server.full.clone();
+        let mut path = self.server().full.clone();
 
         path.path_segments_mut()
-            .map_err(|()| Error::InvalidUrl(self.server.clone()))?
+            .map_err(|()| Error::InvalidUrl(self.server().clone()))?
             .push("lighthouse")
             .push("validator_inclusion")
             .push(&epoch.to_string())
@@ -302,10 +302,10 @@ impl BeaconNodeHttpClient {
     pub async fn get_lighthouse_eth1_syncing(
         &self,
     ) -> Result<GenericResponse<Eth1SyncStatusData>, Error> {
-        let mut path = self.server.full.clone();
+        let mut path = self.server().full.clone();
 
         path.path_segments_mut()
-            .map_err(|()| Error::InvalidUrl(self.server.clone()))?
+            .map_err(|()| Error::InvalidUrl(self.server().clone()))?
             .push("lighthouse")
             .push("eth1")
             .push("syncing");
@@ -317,10 +317,10 @@ impl BeaconNodeHttpClient {
     pub async fn get_lighthouse_eth1_block_cache(
         &self,
     ) -> Result<GenericResponse<Vec<Eth1Block>>, Error> {
-        let mut path = self.server.full.clone();
+        let mut path = self.server().full.clone();
 
         path.path_segments_mut()
-            .map_err(|()| Error::InvalidUrl(self.server.clone()))?
+            .map_err(|()| Error::InvalidUrl(self.server().clone()))?
             .push("lighthouse")
             .push("eth1")
             .push("block_cache");
@@ -332,10 +332,10 @@ impl BeaconNodeHttpClient {
     pub async fn get_lighthouse_eth1_deposit_cache(
         &self,
     ) -> Result<GenericResponse<Vec<DepositLog>>, Error> {
-        let mut path = self.server.full.clone();
+        let mut path = self.server().full.clone();
 
         path.path_segments_mut()
-            .map_err(|()| Error::InvalidUrl(self.server.clone()))?
+            .map_err(|()| Error::InvalidUrl(self.server().clone()))?
             .push("lighthouse")
             .push("eth1")
             .push("deposit_cache");
@@ -345,10 +345,10 @@ impl BeaconNodeHttpClient {
 
     /// `GET lighthouse/staking`
     pub async fn get_lighthouse_staking(&self) -> Result<bool, Error> {
-        let mut path = self.server.full.clone();
+        let mut path = self.server().full.clone();
 
         path.path_segments_mut()
-            .map_err(|()| Error::InvalidUrl(self.server.clone()))?
+            .map_err(|()| Error::InvalidUrl(self.server().clone()))?
             .push("lighthouse")
             .push("staking");
 
@@ -357,10 +357,10 @@ impl BeaconNodeHttpClient {
 
     /// `POST lighthouse/database/reconstruct`
     pub async fn post_lighthouse_database_reconstruct(&self) -> Result<String, Error> {
-        let mut path = self.server.full.clone();
+        let mut path = self.server().full.clone();
 
         path.path_segments_mut()
-            .map_err(|()| Error::InvalidUrl(self.server.clone()))?
+            .map_err(|()| Error::InvalidUrl(self.server().clone()))?
             .push("lighthouse")
             .push("database")
             .push("reconstruct");
@@ -370,10 +370,10 @@ impl BeaconNodeHttpClient {
 
     /// `POST lighthouse/add_peer`
     pub async fn post_lighthouse_add_peer(&self, req: AdminPeer) -> Result<(), Error> {
-        let mut path = self.server.full.clone();
+        let mut path = self.server().full.clone();
 
         path.path_segments_mut()
-            .map_err(|()| Error::InvalidUrl(self.server.clone()))?
+            .map_err(|()| Error::InvalidUrl(self.server().clone()))?
             .push("lighthouse")
             .push("add_peer");
 
@@ -382,10 +382,10 @@ impl BeaconNodeHttpClient {
 
     /// `POST lighthouse/remove_peer`
     pub async fn post_lighthouse_remove_peer(&self, req: AdminPeer) -> Result<(), Error> {
-        let mut path = self.server.full.clone();
+        let mut path = self.server().full.clone();
 
         path.path_segments_mut()
-            .map_err(|()| Error::InvalidUrl(self.server.clone()))?
+            .map_err(|()| Error::InvalidUrl(self.server().clone()))?
             .push("lighthouse")
             .push("remove_peer");
 
@@ -402,10 +402,10 @@ impl BeaconNodeHttpClient {
         start_slot: Slot,
         end_slot: Slot,
     ) -> Result<Vec<BlockReward>, Error> {
-        let mut path = self.server.full.clone();
+        let mut path = self.server().full.clone();
 
         path.path_segments_mut()
-            .map_err(|()| Error::InvalidUrl(self.server.clone()))?
+            .map_err(|()| Error::InvalidUrl(self.server().clone()))?
             .push("lighthouse")
             .push("analysis")
             .push("block_rewards");
@@ -423,10 +423,10 @@ impl BeaconNodeHttpClient {
         start_epoch: Epoch,
         end_epoch: Epoch,
     ) -> Result<Vec<BlockPackingEfficiency>, Error> {
-        let mut path = self.server.full.clone();
+        let mut path = self.server().full.clone();
 
         path.path_segments_mut()
-            .map_err(|()| Error::InvalidUrl(self.server.clone()))?
+            .map_err(|()| Error::InvalidUrl(self.server().clone()))?
             .push("lighthouse")
             .push("analysis")
             .push("block_packing_efficiency");
@@ -445,10 +445,10 @@ impl BeaconNodeHttpClient {
         end_epoch: Epoch,
         target: String,
     ) -> Result<Vec<AttestationPerformance>, Error> {
-        let mut path = self.server.full.clone();
+        let mut path = self.server().full.clone();
 
         path.path_segments_mut()
-            .map_err(|()| Error::InvalidUrl(self.server.clone()))?
+            .map_err(|()| Error::InvalidUrl(self.server().clone()))?
             .push("lighthouse")
             .push("analysis")
             .push("attestation_performance")
